backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
sqlmodel==0.0.21
psycopg[binary]==3.2.8
alembic==1.13.2
redis==5.0.7
celery==5.4.0
python-dotenv==1.0.1
httpx==0.27.2
boto3==1.34.162
python-multipart==0.0.9
stripe==8.10.0
